requirements.txt
streamlit
google-generativeai
langchain
langchain-pinecone
langchain-google-genai
pinecone-client
python-dotenv
